% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{all_data}
\alias{all_data}
\title{All data Algorithm}
\usage{
all_data(
  M,
  Tr,
  rank,
  reference_P = NULL,
  seed = NULL,
  nrun = 5,
  method = "brunet",
  force_second = NULL
)
}
\arguments{
\item{force_second}{}
}
\description{
All data Algorithm
}
