% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_factors}
\alias{assign_factors}
\title{Assign factors to a reference based on cosine similarities}
\usage{
assign_factors(sim_mat, keep_all = FALSE)
}
\arguments{
\item{sim_mat}{similarity matrix between learned and reference factors}
}
\value{
matrix
}
\description{
Assign factors to a reference based on cosine similarities
}
