% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{bootstrap_wrapper}
\alias{bootstrap_wrapper}
\title{Bootstrap wrapper for causalLFO}
\usage{
bootstrap_wrapper(
  algorithm,
  filename,
  M,
  Tr,
  rank,
  seed = NULL,
  nrun = 5,
  method = "brunet",
  reference_P = NULL,
  bootstrap_reps = 500,
  ...
)
}
\arguments{
\item{algorithm}{function, a \code{causalLFO} algorithm (e.g., \link{random_split})}

\item{filename}{string, file name to save bootstrap estimates and aligned
factor matrices}

\item{M}{\code{DxN} matrix, observed outcomes for \code{N} samples}

\item{Tr}{boolean vector of length \code{N}, treatment assignment for \code{N} samples}

\item{rank}{integer, latent rank \code{K}}

\item{reference_P}{\code{DxK} matrix, optional latent factor reference matrix.
Results will be aligned to these factors if provided.}

\item{bootstrap_reps}{integer, number of bootstrap samples to use; default is \code{500}}

\item{...}{other parameters to be passed to \code{algorithm}, e.g., \code{"prop"}
for the \link{random_split} algorithm}
}
\value{
A \code{list} with the following elements:
\itemize{
  \item \code{mean}: bootstrap mean ATE, vector of length \code{K = rank}
  \item \code{se}: bootstrap standard error, vector of length \code{K = rank}
  \item \code{lower95}: lower end of 95\% confidence interval via bootstrap quantiles, vector of length \code{K = rank}
  \item \code{upper95}: upper end of 95\% confidence interval via bootstrap quantiles, vector of length \code{K = rank}
  \item \code{Phat}: consensus factor matrix, elementwise mean across aligned bootstrap replicates
  \item \code{Chat}: consensus factor loadings matrix, generated via NNLM on \code{Phat} and \code{M}
  \item \code{est_file}: string, path to file containing all ATE estimates
  \item \code{all_Ps_file}: string, path to file containing all aligned factor matrices
}
}
\description{
Performs bootstrap resampling and applies the provided algorithm to each
sample. Aligns estimated factor matrices and returns bootsrapped mean, se, and 95% confidence
interval for the ATE. Called internally by causalLFO algorithms (\link{all_data}, \link{random_split},
\link{impute}, \link{stabilize}, and \link{impute_and_stabilize}) when \code{bootstrap = TRUE}.
}
