% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_mat}
\alias{combine_mat}
\title{Combine potential outcomes matrices according to a treatment program}
\usage{
combine_mat(M0, M1, Tr)
}
\arguments{
\item{M0}{matrix DxN, untreated potential outcome}

\item{M1}{matrix DxN, treated potential outcomes}

\item{Tr}{vector length N, binary treatment program assigning each
element 1,...,N to treated (Tr[i] = 1) or untreated (Tr[i] = 0)}
}
\value{
matrix DxN, M combines columns of M0 and M1 according to treatment
program Tr, such that M[,i] = M0[,i] \* (1 - Tr[i]) + M1[,i] \* Tr[i]
}
\description{
Combine potential outcomes matrices according to a treatment program
}
