% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_nmf_info}
\alias{extract_nmf_info}
\title{Extract and process NMF estimates}
\usage{
extract_nmf_info(nmf_res, reference_P = NULL)
}
\arguments{
\item{nmf_res}{result from NMF::nmf}

\item{reference_P}{matrix DxK, reference factor matrix, if available}
}
\value{
list with estimated factors (P), factor loadings (C), and alignment
to reference_P if provided
}
\description{
Extract and process NMF estimates
}
