% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{impute}
\alias{impute}
\title{Impute algorithm}
\usage{
impute(
  M,
  Tr,
  rank,
  reference_P = NULL,
  seed = NULL,
  nrun = 5,
  method = "brunet",
  force_second = NULL
)
}
\arguments{
\item{force_second}{}
}
\description{
Impute algorithm
}
