% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{impute_and_stabilize}
\alias{impute_and_stabilize}
\title{Impute and Stabilize algorithm}
\usage{
impute_and_stabilize(
  M,
  Tr,
  rank,
  reference_P = NULL,
  seed = NULL,
  nrun = 5,
  method = "brunet",
  force_second = NULL
)
}
\arguments{
\item{force_second}{}
}
\description{
Impute and Stabilize algorithm
}
