% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nmf_wrapper}
\alias{nmf_wrapper}
\title{NMF wrapper}
\usage{
nmf_wrapper(M, rank, nrun = 5, seed = NULL, method = "brunet", ...)
}
\arguments{
\item{M}{matrix DxN, NMF input matrix}

\item{rank}{integer K << D, N, latent rank}

\item{nrun}{integer, number of NMF runs (passed to NMF::nmf)}

\item{seed}{integer, random seed (passed to NMF::nmf)}

\item{method}{string, NMF method (passed to NMF::nmf)}
}
\value{
NMF result, value from NMF::nmf
}
\description{
NMF wrapper
}
