% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot.causalLFO_bootstrap_result}
\alias{plot.causalLFO_bootstrap_result}
\title{Plot a causalLFO_bootstrap_result object}
\usage{
\method{plot}{causalLFO_bootstrap_result}(x)
}
\arguments{
\item{object}{causalLFO_result object}
}
\description{
Plot a causalLFO_bootstrap_result object
}
