% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_causalLFO_bootstrap_results}
\alias{plot_causalLFO_bootstrap_results}
\title{Plot multiple causalLFO_bootstrap_result objects}
\usage{
plot_causalLFO_bootstrap_results(res_list, group_name = "")
}
\arguments{
\item{res_list}{named list of causalLFO_bootstrap_result objects. Names will become colors}

\item{group_name}{string, name of grouping variable / legend title for colors}
}
\value{
ggplot2 object
}
\description{
Plot multiple causalLFO_bootstrap_result objects
}
