% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{poisson_nnlm}
\alias{poisson_nnlm}
\title{Poisson-likelihood non-negative linear model with fixed factor matrix using
Lee and Seung 1999-style multiplicative updates}
\usage{
poisson_nnlm(M, fixed_P, maxiter = 10000, tol = 1e-06)
}
\arguments{
\item{M}{matrix DxN, data matrix}

\item{fixed_P}{matrix DxK, fixed factor matrix}

\item{maxiter}{int, max number of gradient descent updates}

\item{tol}{float, tolerance for convergence}
}
\value{
C, matrix KxN, factor loadings matrix
}
\description{
Poisson-likelihood non-negative linear model with fixed factor matrix using
Lee and Seung 1999-style multiplicative updates
}
