% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{random_split}
\alias{random_split}
\title{Random split algorithm}
\usage{
random_split(
  M,
  Tr,
  rank,
  reference_P = NULL,
  prop = 0.5,
  seed = NULL,
  nrun = 5,
  method = "brunet",
  force_second = c()
)
}
\arguments{
\item{force_second}{}
}
\description{
Random split algorithm
}
