% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{stabilize}
\alias{stabilize}
\title{Stabilize algorithm}
\usage{
stabilize(
  M,
  Tr,
  rank,
  reference_P = NULL,
  seed = NULL,
  nrun = 5,
  method = "brunet",
  force_second = NULL
)
}
\arguments{
\item{force_second}{}
}
\description{
Stabilize algorithm
}
