% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summary.causalLFO_result}
\alias{summary.causalLFO_result}
\title{Summary of a causalLFO_result object}
\usage{
\method{summary}{causalLFO_result}(object)
}
\arguments{
\item{object}{causalLFO_result object}
}
\description{
Summary of a causalLFO_result object
}
